backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]  # Para JWT
passlib[bcrypt]             # Para encriptar contraseñas
python-dotenv               # Para leer variables de entorno
